move the update username route to v3
